Player implementation
-expanded on constructor
-implemented armored and health

+++ src/entity/Armored.java
@@ -1,6 +1,6 @@
 package entity;
 
 public interface Armored {
-	public double getFlatArmor();
-	public double getProportionalArmor();
+	public int getFlatArmor();
+	public int getProportionalArmor();
 }
